x/xarchain/types/message_update_intent_test.go
package types

import (
	"testing"

	"xarchain/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgUpdateIntent_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateIntent
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateIntent{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateIntent{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
